Added changes from HPL 2.0: 64-bit PRNG, better residual check, fixes to address calculation on 64-bit systems.

+++ hpl/src/auxil/HPL_dlange.c
@@ -1,10 +1,10 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 2.0 - September 10, 2008                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
- *    Innovative Computing Laboratories                                 
- *    (C) Copyright 2000-2004 All Rights Reserved                       
+ *    Innovative Computing Laboratory                                 
+ *    (C) Copyright 2000-2008 All Rights Reserved                       
  *                                                                      
  * -- Copyright notice and Licensing terms:                             
  *                                                                      
@@ -22,7 +22,7 @@
  * 3. All  advertising  materials  mentioning  features  or  use of this
  * software must display the following acknowledgement:                 
  * This  product  includes  software  developed  at  the  University  of
- * Tennessee, Knoxville, Innovative Computing Laboratories.             
+ * Tennessee, Knoxville, Innovative Computing Laboratory.             
  *                                                                      
  * 4. The name of the  University,  the name of the  Laboratory,  or the
  * names  of  its  contributors  may  not  be used to endorse or promote
@@ -134,7 +134,7 @@ double HPL_dlange
 /*
  * Find norm_1( A ).
  */
-      work = (double*)malloc( N * sizeof( double ) );
+      work = (double*)malloc( (size_t)(N) * sizeof( double ) );
       if( work == NULL )
       { HPL_abort( __LINE__, "HPL_dlange", "Memory allocation failed" ); }
       else
@@ -157,7 +157,7 @@ double HPL_dlange
 /*
  * Find norm_inf( A )
  */
-      work = (double*)malloc( M * sizeof( double ) );
+      work = (double*)malloc( (size_t)(M) * sizeof( double ) );
       if( work == NULL )
       { HPL_abort( __LINE__, "HPL_dlange", "Memory allocation failed" ); }
       else
