+++ hpl/src/pauxil/HPL_dlaswp02N.c
@@ -1,10 +1,10 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 2.0 - September 10, 2008                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
- *    Innovative Computing Laboratories                                 
- *    (C) Copyright 2000-2004 All Rights Reserved                       
+ *    Innovative Computing Laboratory                                 
+ *    (C) Copyright 2000-2008 All Rights Reserved                       
  *                                                                      
  * -- Copyright notice and Licensing terms:                             
  *                                                                      
@@ -22,7 +22,7 @@
  * 3. All  advertising  materials  mentioning  features  or  use of this
  * software must display the following acknowledgement:                 
  * This  product  includes  software  developed  at  the  University  of
- * Tennessee, Knoxville, Innovative Computing Laboratories.             
+ * Tennessee, Knoxville, Innovative Computing Laboratory.             
  *                                                                      
  * 4. The name of the  University,  the name of the  Laboratory,  or the
  * names  of  its  contributors  may  not  be used to endorse or promote
@@ -148,7 +148,8 @@ void HPL_dlaswp02N
  */
    if( ( M <= 0 ) || ( N <= 0 ) ) return;
 
-   for( i = 0; i < M; i++ ) *(W0+i*LDW) = (double)(LINDXAU[i]);
+   for( i = 0; i < M; i++ ) 
+      *(W0+(size_t)(i)*(size_t)(LDW)) = (double)(LINDXAU[i]);
 
    nr = N - ( nu = (int)( ( (unsigned int)(N) >> HPL_LASWP02N_LOG2_DEPTH ) <<
                           HPL_LASWP02N_LOG2_DEPTH ) );
@@ -158,7 +159,7 @@ void HPL_dlaswp02N
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A0 + LINDXA[i]; w0 = W + i * LDW;
+         a0 = A0 + (size_t)(LINDXA[i]); w0 = W + (size_t)(i) * (size_t)(LDW);
 
          w0[ 0] = *a0; a0 += LDA;
 #if ( HPL_LASWP02N_DEPTH >  1 )
@@ -194,7 +195,7 @@ void HPL_dlaswp02N
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A0 + LINDXA[i]; w0 = W + i * LDW;
+         a0 = A0 + (size_t)(LINDXA[i]); w0 = W + (size_t)(i) * (size_t)(LDW);
          for( j = 0; j < nr; j++, a0 += LDA ) { w0[j] = *a0; }
       }
    }
