+++ hpl/src/pfact/HPL_dlocswpN.c
@@ -1,10 +1,10 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 2.0 - September 10, 2008                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
- *    Innovative Computing Laboratories                                 
- *    (C) Copyright 2000-2004 All Rights Reserved                       
+ *    Innovative Computing Laboratory                                 
+ *    (C) Copyright 2000-2008 All Rights Reserved                       
  *                                                                      
  * -- Copyright notice and Licensing terms:                             
  *                                                                      
@@ -22,7 +22,7 @@
  * 3. All  advertising  materials  mentioning  features  or  use of this
  * software must display the following acknowledgement:                 
  * This  product  includes  software  developed  at  the  University  of
- * Tennessee, Knoxville, Innovative Computing Laboratories.             
+ * Tennessee, Knoxville, Innovative Computing Laboratory.             
  *                                                                      
  * 4. The name of the  University,  the name of the  Laboratory,  or the
  * names  of  its  contributors  may  not  be used to endorse or promote
@@ -347,7 +347,7 @@ void HPL_dlocswpN
  */
          if( myrow == (int)(WORK[3]) )
          {
-            A2 = Mptr( PANEL->A, II + (int)(WORK[1]), 0, lda );
+            A2 = Mptr( PANEL->A, II + (size_t)(WORK[1]), 0, lda );
 
             for( i = 0; i < nu; i += HPL_LOCSWP_DEPTH,
                  Wr0 += HPL_LOCSWP_DEPTH )
